feat: Ability to specify values for attrivutes Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
